Will add Grid container to html and squares
Took a break for a couple of days because of Apex, but also
because I couldn't figure it out. Reached out to TOP for help on a basic
question. Created a prompt to ask how many columns/rows.
Works in terms of answering the question but currently only coding
to make this work for uniform grids (2x2, 4x4, 8x8). Once i figure it
out I will altercode to make it work for non uniform grids

Was stuck intially as my create grid function was skipping over the
for loop I created and I didn't know why. Turns out I had to actually
run the function. Only confused as previously I don't believe I did, but
it makes sense. Once I did that I successfully created code that made
multiple square divs. and separted them by adding i to the id string to
identify each as needed