Removed second EventHandler + CSS info
Turns out I didn't need that second set of EventHandlers
for the ResetGrid function. Maybe I entered something wrong before
but it wasn't making sense to me. I removed it and code still
ran fine.

Forgot to include CSS edits
Set Body to Flex:
Flex Direction to Column to stack the reset button above the grid
Justified and Aligned everythign to the center.
Added a gap of 5px between button and grid
In order to everything I had to set the height of the body to the
height of the viewport, then give it a margin of 0 px;